hw4/requirements.txt
wheel
flask
google-cloud-datastore
google-cloud-firestore
gunicorn
rsa